requirements.txt
requests
keyboard
pynput
pyyaml
